requirements.txt
absl-py==0.2.0
astor==0.6.2
bleach==1.5.0
certifi==2018.4.16
chardet==3.0.4
future==0.16.0
gast==0.2.0
grpcio==1.11.0
gym==0.10.5
html5lib==0.9999999
idna==2.6
Markdown==2.6.11
numpy==1.14.3
protobuf==3.5.2.post1
pygame==1.9.3
pyglet==1.3.2
requests==2.18.4
six==1.11.0
tensorboard==1.8.0
tensorflow-gpu==1.8.0
termcolor==1.1.0
urllib3==1.22
Werkzeug==0.14.1
